fixed some types problem for generic IDs

+++ db_managers/simple_movie_db_manager/src/movie_item.rs
@@ -25,8 +25,8 @@ pub struct MovieItem {
 impl Item for MovieItem {
     type ID = i32;
     
-    fn id(&self) -> u64 {
-        self.id as u64
+    fn id(&self) -> i32 {
+        self.id
     }
     fn name(&self) -> String {
         self.name.clone()
